+++ source/Relativity.Testing.Framework.Api/Strategies/KeyboardShortcuts/KeyboardShortcutsGetStrategyNotSupported.cs
@@ -9,9 +9,14 @@ namespace Relativity.Testing.Framework.Api.Strategies.KeyboardShortcuts
 	[VersionRange("<12.1")]
 	internal class KeyboardShortcutsGetStrategyNotSupported : IKeyboardShortcutsGetStrategy
 	{
-		public Task<IEnumerable<KeyboardShortcut>> GetKeyboardShortcutsAsync(int workspaceId, KeyboardShortcutsIncludeOptions includeOptions = null)
+		public IEnumerable<KeyboardShortcut> Get(int workspaceId, KeyboardShortcutsIncludeOptions includeOptions = null)
 		{
-			throw new ArgumentException("The method GetKeyboardShortcuts does not support version of Relativity lower than 12.1.");
+			throw new ArgumentException("The method Get for KeyboardShortcuts does not support version of Relativity lower than 12.1.");
+		}
+
+		public Task<IEnumerable<KeyboardShortcut>> GetAsync(int workspaceId, KeyboardShortcutsIncludeOptions includeOptions = null)
+		{
+			throw new ArgumentException("The method GetAsync fro KeyboardShortcuts does not support version of Relativity lower than 12.1.");
 		}
 	}
 }